# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Serviços Assinatura",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/laravel-api-swagger/public/api"
        }
    ],
    "paths": {
        "/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Listagem de Posts",
                "operationId": "App\\Http\\Controllers\\PostController::index",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Cadastro de posts",
                "operationId": "App\\Http\\Controllers\\PostController::store",
                "requestBody": {
                    "description": "Dados para requisição",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/PostRequisicao"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post salvo com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Falha de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erro"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/posts/{cd_post}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Lista um post pelo código",
                "operationId": "App\\Http\\Controllers\\PostController::show",
                "parameters": [
                    {
                        "name": "cd_post",
                        "in": "path",
                        "description": "Código do post a ser listado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erro"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Atualizar um post pelo código",
                "operationId": "App\\Http\\Controllers\\PostController::update",
                "parameters": [
                    {
                        "name": "cd_post",
                        "in": "path",
                        "description": "Código do post a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados para requisição",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/PostRequisicao"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erro"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Falha de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erro"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Deletar um post pelo código",
                "operationId": "App\\Http\\Controllers\\PostController::destroy",
                "parameters": [
                    {
                        "name": "cd_post",
                        "in": "path",
                        "description": "Código do post a ser deletado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Post deletado com sucesso"
                    },
                    "404": {
                        "description": "Post não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erro"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Erro": {
                "properties": {
                    "error": {
                        "description": "Mensagem de erro/validação.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PostRequisicao": {
                "required": [
                    "nm_post",
                    "ds_post"
                ],
                "properties": {
                    "nm_post": {
                        "description": "Nome do post",
                        "type": "string"
                    },
                    "ds_post": {
                        "description": "Descrição do post",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Post": {
                "properties": {
                    "cd_post": {
                        "description": "Código do post",
                        "type": "integer"
                    },
                    "nm_post": {
                        "description": "Nome do post",
                        "type": "string"
                    },
                    "ds_post": {
                        "description": "Descrição do post",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Data de criação do Post",
                        "type": "string"
                    },
                    "updated_at": {
                        "description": "Data de atualização do Post",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}